Adapt unittest with the new search by ancestors
