+++ base/business/education_group.py
@@ -25,10 +25,11 @@
 ##############################################################################
 from django.core.exceptions import PermissionDenied
 
-from base.models import offer_year_entity
+from base.models.entity import Entity
 from base.models.enums import offer_year_entity_type
 from base.models.person import Person
 from base.models.program_manager import is_program_manager
+from base.models.person_entity import is_attached_entities
 
 
 def can_user_edit_administrative_data(a_user, an_education_group_year):
@@ -48,11 +49,10 @@ def can_user_edit_administrative_data(a_user, an_education_group_year):
 
 
 def _is_management_entity_linked_to_user(person, an_education_group_year):
-    management_entity_ids = list(offer_year_entity.find_by_education_group_year(
-        education_group_yr=an_education_group_year,
-        typ=offer_year_entity_type.ENTITY_MANAGEMENT).values_list('entity', flat=True))
+    entities = Entity.objects.filter(offeryearentity__education_group_year=an_education_group_year,
+                                     offeryearentity__type=offer_year_entity_type.ENTITY_MANAGEMENT)
 
-    return any(entity.id in management_entity_ids for entity in person.entities)
+    return is_attached_entities(person, entities)
 
 
 def assert_category_of_education_group_year(education_group_year, authorized_categories):
