+++ base/models/person.py
@@ -24,21 +24,21 @@
 #
 ##############################################################################
 from datetime import date
+
+from django.conf import settings
+from django.contrib.auth.models import User
 from django.db import models
 from django.db.models import Q
-from django.contrib.auth.models import User
+from django.db.models import Value
+from django.db.models.functions import Concat, Lower
 from django.utils.functional import cached_property
 from django.utils.translation import ugettext_lazy as _
-from django.conf import settings
 
 from base.business.learning_units.perms import is_person_linked_to_entity_in_charge_of_learning_unit
 from base.models.entity_version import find_main_entities_version
-from base.models.person_entity import find_entities_by_person
+from base.models.enums import person_source_type
 from base.models.utils.person_entity_filter import filter_by_attached_entities
 from osis_common.models.serializable_model import SerializableModel, SerializableModelAdmin
-from base.models.enums import person_source_type
-from django.db.models import Value
-from django.db.models.functions import Concat, Lower
 
 CENTRAL_MANAGER_GROUP = "central_managers"
 FACULTY_MANAGER_GROUP = "faculty_managers"
@@ -142,10 +142,6 @@ class Person(SerializableModel):
     @cached_property
     def find_main_entities_version(self):
         return filter_by_attached_entities(self, find_main_entities_version())
-
-    @cached_property
-    def entities(self):
-        return find_entities_by_person(self)
 
 
 def find_by_id(person_id):
